Added some more profiling logging.
Added the filename of the file under analysis.
Added the number of classified items.
Both info are logged at the beginning and at the end of the program.
This should help in handling situations where there is the interleaved
logging of different files, or there is the external editing of the CSV
files.

Meanwhile, some refactoring was done, by introducing the helper
functions in the WordList class.