Splitted lemmatization from n-grams generation.
Now preprocess.py takes as input a CSV file with a column named
'abstract1'.
It generates a CSV with a new column named 'abstract_lem', which
contains the lemmatized version of the abstract.
In this setup, gen-n-grams.py can now be applied to a CSV containing a
column named 'abstract_lem' and produces the list of terms.

In this way, the lemmatized version of the original abstract is
persisted in the CSV file for further processing.