Save the gensim Dictionary object in a file when saving the model
It seems that gensim is not able to properly save and load the
dictionary object when saving the model.
Without the Dictionary object it is not possible to calculate the
coherence from a loaded model.
With this trick we can properly load the dictionary object with the
model and compute the coherence scores.