Reworked the logic to replace acronyms in the preprocessing stage
Instead of using an inefficient set of loop to go through the acronyms,
now a dict is used.
Moreover, instead of checking all the acronyms, which is definitely
inefficient for a long list of acronyms, now the acronyms for each text
are derived using the same algorithm that is used in the "acronyms"
stage.
Therefore, only the acronyms appearing the abstracts are checked.